requirements.txt
matplotlib
librosa
numpy
progressbar2
pydub
requests
skeltorch
srt
torch
torchaudio